Refine, modularize granular related genes analytics
